:sparkles: feat(animal): historial de peso del animal (WeightRecord)
El peso del animal deja de ser un escalar y pasa a ser una serie temporal. Nuevo agregado WeightRecord (dominio, puertos in/out, casos de uso crear/eliminar/listar/último, JPA + mapper + repos, controller y DTOs) y migración 162 que crea `weight_records`.

- Animal.weight pasa a nullable; el peso "actual" se deriva del último registro (weightMeasuredAt en la respuesta).
- Nueva WeightSource, AnimalRef y WeightRecordNotFoundException (registrada en GlobalExceptionHandler → 404).

+++ src/main/java/com/vetsoftware/app/animal/application/command/CreateAnimalCommand.java
@@ -4,11 +4,12 @@ import com.vetsoftware.app.animal.domain.AnimalType;
 import com.vetsoftware.app.animal.domain.Gender;
 import com.vetsoftware.app.animal.domain.ReproductiveState;
 import com.vetsoftware.app.animal.domain.WeightType;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record CreateAnimalCommand(
         String name, String code, Long specieId, Long breedId, Long ownerId,
         Gender gender, WeightType weightType, AnimalType animalType,
         ReproductiveState reproductiveState, Long colorId, LocalDate bod,
-        Integer weight, Integer size, boolean deceased, LocalDate deceasedDate, Long companyId
+        BigDecimal weight, Integer size, boolean deceased, LocalDate deceasedDate, Long companyId
 ) {}
